Chore: Evaluate UserItem to Typescript and moved from presentation to container
