+++ app/containers/UnreadBadge/index.tsx
@@ -3,7 +3,7 @@ import { StyleSheet, Text, View, ViewStyle } from 'react-native';
 
 import sharedStyles from '../../views/Styles';
 import { getUnreadStyle } from './getUnreadStyle';
-import { withTheme } from '../../theme';
+import { useTheme } from '../../theme';
 
 const styles = StyleSheet.create({
 	unreadNumberContainerNormal: {
@@ -29,8 +29,7 @@ const styles = StyleSheet.create({
 	}
 });
 
-interface IUnreadBadge {
-	theme?: string;
+export interface IUnreadBadge {
 	unread?: number;
 	userMentions?: number;
 	groupMentions?: number;
@@ -42,7 +41,9 @@ interface IUnreadBadge {
 }
 
 const UnreadBadge = React.memo(
-	({ theme, unread, userMentions, groupMentions, style, tunread, tunreadUser, tunreadGroup, small }: IUnreadBadge) => {
+	({ unread, userMentions, groupMentions, style, tunread, tunreadUser, tunreadGroup, small }: IUnreadBadge) => {
+		const { theme } = useTheme();
+
 		if ((!unread || unread <= 0) && !tunread?.length) {
 			return null;
 		}
@@ -88,4 +89,4 @@ const UnreadBadge = React.memo(
 	}
 );
 
-export default withTheme(UnreadBadge);
+export default UnreadBadge;
