+++ app/ee/omnichannel/containers/OmnichannelStatus.tsx
@@ -5,7 +5,7 @@ import * as List from '../../../containers/List';
 import styles from '../../../views/RoomsListView/styles';
 import { SWITCH_TRACK_COLOR, themes } from '../../../constants/colors';
 import { useTheme } from '../../../theme';
-import UnreadBadge from '../../../presentation/UnreadBadge';
+import UnreadBadge from '../../../containers/UnreadBadge';
 import RocketChat from '../../../lib/rocketchat';
 import { changeLivechatStatus, isOmnichannelStatusAvailable } from '../lib';
 import { IUser } from '../../../definitions/IUser';
